# openapi.yaml
openapi: 3.0.3
info:
  title: Releasarr API
  version: 0.1.0
  description: |
    Contract derived from the new frontend client. Captures the HTTP interface the
    UI expects today so the backend can be validated and mocked consistently.
  license:
    name: Proprietary
    url: https://github.com/koreetz/releasarr
servers:
  - url: https://api.releasarr.dev/api
    description: Production server
  - url: http://localhost:8001/api
    description: Local development server
security:
  - BearerAuth: []
  - ServiceKeyAuth: []
paths:
  /requests:
    get:
      tags: [Requests]
      summary: List media requests
      operationId: listRequests
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/RequestStatus'
        - $ref: '#/components/parameters/RequestType'
        - $ref: '#/components/parameters/RequestOwner'
      responses:
        '200':
          description: Paged list of media requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestsResponse'
        '400':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The caller is restricted to their own requests and may not filter by owner.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Requests]
      summary: Create a media request
      operationId: createRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaRequestCreate'
      responses:
        '201':
          description: Created request details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaRequest'
        '400':
          description: Malformed request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /discover/search:
    get:
      tags: [Discover]
      summary: Search TVDB or TMDB for media to request
      description: |
        Results come from the metadata provider for the given type - TVDB for
        series, TMDB for movies - and are annotated with what releasarr already
        holds: whether Sonarr/Radarr has the item in its library, and which of
        its seasons already have a request. The annotation is best-effort, so a
        Sonarr or Radarr that cannot be reached leaves the results unannotated
        rather than failing the search.

        Results are ordered by how closely the term matches any title the
        provider knows for an entry - not only the localized one shown - and
        then by how widely known the entry is, so a series outranks the talk
        shows, parodies and featurettes named after it.

        Omitting `type` searches both providers and returns one mixed list. Only
        one of the two providers has to be configured, and a provider that
        cannot be reached is left out rather than failing the search.
      operationId: searchMedia
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            minLength: 1
          description: Free-text title to search for.
        - in: query
          name: type
          required: false
          schema:
            $ref: '#/components/schemas/MediaType'
          description: |
            Restrict the search to one media type. Both are searched when
            omitted.
        - in: query
          name: lang
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 8
          description: |
            Preferred language for the titles and overviews, as a 2- or
            3-letter code. Takes precedence over the configured metadata
            languages, and is ignored if the provider has no translation.
      responses:
        '200':
          description: Metadata provider matches, annotated with library and request state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaSearchResponse'
        '422':
          description: Invalid search parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr, Radarr or the metadata provider could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: The metadata provider is not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /discover/series/{tvdbId}/seasons:
    get:
      tags: [Discover]
      summary: List the seasons of a series, with their request state
      description: |
        Season numbers come from Sonarr's own lookup, so they are the ones an add
        would actually monitor. For a series already in the library each season
        also reports whether Sonarr monitors it and whether it already has a
        request.
      operationId: listSeriesSeasons
      parameters:
        - $ref: '#/components/parameters/TvdbId'
      responses:
        '200':
          description: The seasons the series offers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeriesSeasonsResponse'
        '404':
          description: No series matches the TVDB id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr or the metadata provider could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /discover/root-folders:
    get:
      tags: [Discover]
      summary: List the library locations media can be added to
      operationId: listRootFolders
      parameters:
        - in: query
          name: type
          required: true
          schema:
            $ref: '#/components/schemas/MediaType'
          description: Whether to read Sonarr's or Radarr's root folders.
      responses:
        '200':
          description: Root folders configured in the *arr app, unreachable ones omitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RootFoldersResponse'
        '422':
          description: Invalid media type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr or Radarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /discover/requests:
    post:
      tags: [Discover]
      summary: Add media to Sonarr/Radarr and request it
      description: |
        Adds the picked series or movie to Sonarr/Radarr by its TVDB/TMDB id,
        monitoring the selected seasons, and creates the matching media requests
        before responding. Media already in the library is not re-added: only the
        monitoring is widened to cover the selection.
      operationId: addRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRequestPayload'
      responses:
        '201':
          description: The created requests, one per season for a series.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddRequestResponse'
        '400':
          description: Invalid root folder, season selection or quality profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No media matches the provider id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed for the provided fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Sonarr or Radarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /logs:
    get:
      tags: [Logs]
      summary: List processing logs
      description: |
        Application log entries, newest first. Filters match fields the log
        producer bound onto the record, so `task` returns everything logged
        while that background task was running, including lines emitted deeper
        in the call stack, and `service` splits the API's records from the
        scheduler's. `min_level` is a threshold over the parsed severity.
      operationId: listRequestLogs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/RequestIdFilter'
        - $ref: '#/components/parameters/TaskFilter'
        - $ref: '#/components/parameters/ServiceFilter'
        - $ref: '#/components/parameters/LogLevelFilter'
      responses:
        '200':
          description: Request logs matching the supplied filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        '400':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /indexers/logs:
    get:
      tags: [Indexers]
      summary: List Prowlarr's own log
      description: |
        Prowlarr's application log, newest first — what its own UI shows under
        System → Events. This is Prowlarr's log rather than releasarr's, so it
        is where an indexer's connection failures and captcha trouble surface.
      operationId: listIndexerLogs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/IndexerLogLevelFilter'
      responses:
        '200':
          description: Log entries matching the supplied filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexerLogsResponse'
        '422':
          description: Invalid pagination or filter parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Prowlarr could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Prowlarr is not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token obtained from /auth/login or /auth/refresh.
    ServiceKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        A service API key, acting as the user it is bound to. May be combined
        with X-Act-As-User to act as a different user, when the key allows it.
  parameters:
...
    TvdbId:
      in: path
      name: tvdbId
      required: true
      schema:
        type: integer
        minimum: 1
      description: TVDB identifier of a series.
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
      description: Results page to retrieve (1-indexed).
    PerPage:
      in: query
      name: per_page
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Number of items to return per page.
    RequestStatus:
      in: query
      name: status
      schema:
        $ref: '#/components/schemas/MediaRequestStatus'
      description: Filter requests by lifecycle status.
    RequestType:
      in: query
      name: type
      schema:
        $ref: '#/components/schemas/MediaType'
      description: Filter requests by media type.
    RequestOwner:
      in: query
      name: owner
      schema:
        type: string
      description: >-
        Filter requests by owner user id. Only callers with view_all_requests
        (or an admin) may use this; others are restricted to their own.
...
    RequestIdFilter:
      in: query
      name: request_id
      schema:
        type: string
      description: Optional media request identifier to filter results.
    TaskFilter:
      in: query
      name: task
      schema:
        $ref: '#/components/schemas/SyncJobKind'
      description: Optional background task to filter results.
    ServiceFilter:
      in: query
      name: service
      schema:
        $ref: '#/components/schemas/LogService'
      description: Optional process to filter results by.
    LogLevelFilter:
      in: query
      name: min_level
      schema:
        $ref: '#/components/schemas/RequestLogLevel'
      description: |
        Least severe level to return, along with everything worse. Records are
        mapped onto three levels, so this is a threshold rather than an exact
        match: `warning` also returns errors.
...
    IndexerLogLevelFilter:
      in: query
      name: min_level
      schema:
        $ref: '#/components/schemas/IndexerLogLevel'
      description: |
        Least severe level to return. Prowlarr treats this as a threshold rather
        than an exact match, so `warn` also returns errors and fatals.
  schemas:
    MediaType:
      type: string
      enum: [movie, series]
    MediaRequestStatus:
      type: string
      enum: [pending, searching, downloading, completed, failed]
...
    LogService:
      type: string
      enum: [api, scheduler]
      description: |
        The process that wrote a record. The API and the scheduler run as
        separate processes, each writing its own log file.
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error identifier.
        message:
          type: string
          description: Human-readable explanation of the error.
        details:
          description: Optional key/value data with context for the failure.
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
              - type: array
                items:
                  type: string
    RequestsResponse:
      type: object
      required: [requests, total, page, per_page]
      properties:
        requests:
          type: array
          items:
            $ref: '#/components/schemas/MediaRequest'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
...
    LogsResponse:
      type: object
      required: [logs, total, page, per_page]
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/RequestLogEntry'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
...
    MovieRequest:
      allOf:
        - $ref: '#/components/schemas/BaseMediaRequest'
        - type: object
          required: [type, runtime, imdb_id]
          properties:
            type:
              type: string
              enum: [movie]
            runtime:
              type: integer
              description: Runtime in minutes.
            imdb_id:
              type: string
              description: IMDb identifier.
            radarr_movie_id:
              type: integer
              nullable: true
              description: Identifier of the linked Radarr movie when the request was auto-synchronised.
    SeriesRequest:
      allOf:
        - $ref: '#/components/schemas/BaseMediaRequest'
        - type: object
          required:
            - type
            - season_number
            - total_episodes
            - series_title
            - series_year
            - imdb_id
          properties:
            type:
              type: string
              enum: [series]
            season_number:
              type: integer
            total_episodes:
              type: integer
            series_title:
              type: string
            series_year:
              type: integer
            imdb_id:
              type: string
            sonarr_series_id:
              type: integer
              nullable: true
              description: Identifier of the linked Sonarr series when the request was auto-synchronised.
            episode_counts:
              $ref: '#/components/schemas/SeriesEpisodeCounts'
              nullable: true
              description: Derived counts of downloaded, pending and unaired episodes for series requests.
    MediaRequest:
      oneOf:
        - $ref: '#/components/schemas/MovieRequest'
        - $ref: '#/components/schemas/SeriesRequest'
      discriminator:
        propertyName: type
        mapping:
          movie: '#/components/schemas/MovieRequest'
          series: '#/components/schemas/SeriesRequest'
    CreateMovieRequest:
      type: object
      required: [type, title, year, runtime, imdb_id]
      properties:
        type:
          type: string
          enum: [movie]
        title:
          type: string
        year:
          type: integer
        runtime:
          type: integer
        imdb_id:
          type: string
        overview:
          type: string
        poster_url:
          type: string
        genres:
          type: array
          items:
            type: string
        localizations:
          type: object
          description: Optional localized titles and overviews keyed by language codes.
          additionalProperties:
            $ref: '#/components/schemas/MediaLocalization'
    CreateSeriesRequest:
      type: object
      required:
        - type
        - title
        - year
        - season_number
        - total_episodes
        - series_title
        - series_year
        - imdb_id
      properties:
        type:
          type: string
          enum: [series]
        title:
          type: string
        year:
          type: integer
        season_number:
          type: integer
        total_episodes:
          type: integer
        series_title:
          type: string
        series_year:
          type: integer
        imdb_id:
          type: string
        overview:
          type: string
        poster_url:
          type: string
        genres:
          type: array
          items:
            type: string
        localizations:
          type: object
          description: Optional localized titles and overviews keyed by language codes.
          additionalProperties:
            $ref: '#/components/schemas/MediaLocalization'
    MediaRequestCreate:
      oneOf:
        - $ref: '#/components/schemas/CreateMovieRequest'
        - $ref: '#/components/schemas/CreateSeriesRequest'
      discriminator:
        propertyName: type
        mapping:
          movie: '#/components/schemas/CreateMovieRequest'
          series: '#/components/schemas/CreateSeriesRequest'
...
    IndexerLogsResponse:
      type: object
      required: [logs, total, page, per_page]
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/IndexerLogEntry'
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
...
    MediaSearchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/MediaSearchResult'
...
    SeriesSeasonsResponse:
      type: object
      required: [seasons]
      properties:
        tvdb_id:
          type: integer
          nullable: true
          description: >-
            Absent when the seasons were read from a series in the library rather
            than looked up by TVDB id, which is how managing a request arrives.
        in_library:
          type: boolean
          default: false
        library_id:
          type: integer
          nullable: true
        monitor_new_seasons:
          type: boolean
          default: false
          description: >-
            Whether Sonarr monitors seasons announced after the series was added.
            Only meaningful in the library; Sonarr has nowhere to record it until
            then.
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/SeasonOption'
    RootFolder:
      type: object
      required: [path]
      properties:
        path:
          type: string
        free_space:
          type: integer
          nullable: true
          description: Bytes available, as the *arr app reports them.
    RootFoldersResponse:
      type: object
      required: [folders]
      properties:
        folders:
          type: array
          items:
            $ref: '#/components/schemas/RootFolder'
    AddRequestPayload:
      type: object
      required: [type, provider_id, root_folder_path]
      properties:
        type:
          $ref: '#/components/schemas/MediaType'
        provider_id:
          type: integer
          description: TVDB id for a series, TMDB id for a movie.
        root_folder_path:
          type: string
          minLength: 1
        season_numbers:
          type: array
          nullable: true
          description: >-
            Seasons to request. Required to add a series Sonarr does not hold
            yet, and rejected for movies. Empty for a series already in the
            library requests nothing and sets only monitor_new_seasons.
          items:
            type: integer
        monitor_new_seasons:
          type: boolean
          default: false
          description: >-
            Ask Sonarr to monitor seasons announced later, which the sync then
            turns into requests of their own. Series only.
...
    AddRequestResponse:
      type: object
      required: [requests]
      properties:
        requests:
          type: array
          description: One request per requested season for a series, a single one for a movie.
          items:
            $ref: '#/components/schemas/MediaRequest'
...